+++ src/components/TextField.vue
@@ -11,6 +11,7 @@
             :required="required"
             :aria-required="required"
             :aria-invalid="error"
+            :disabled="disabled"
         />
     </div>
 </template>
@@ -39,6 +40,10 @@ const props = defineProps({
         default: null,
     },
     required: {
+        type: Boolean,
+        default: false,
+    },
+    disabled: {
         type: Boolean,
         default: false,
     },