+++ History.md
@@ -2,6 +2,18 @@
 
 Please register all relevant modifications to files, JS and state if these modifications took place in a staging, production or local environment, to which branches the files were commited and who did those changes.
 
+December 21th, 2023 | Upload to production | Julieta
+
+-   **Description**: Disabled option has been added in all form-fields components. - Upload version 0.0.5
+-   **Main Card**: hthttps://app.clickup.com/t/862kk6p27
+-   **Vue**:
+    -   components/Chackbox.vue
+    -   components/GroupCheckbox.vue
+    -   components/GroupRadio.vue
+    -   components/TextArea.vue
+    -   components/TextField.vue
+    -   components/UploadFile.vue
+
 December 12th, 2023 | Upload to production | Julieta
 
 -   **Description**: UploadFile has been added as a new component and the Error component has been separated from the Hint component. - Upload version 0.0.2